Rename jorb.c as a header instead
